feat: Implement building and resource management features with new map UI components and updated data models.

+++ app/src/main/java/com/example/villageevo/domain/map/MapData.kt
@@ -10,8 +10,8 @@ data class MapMetaData (
 )
 data class PotentialData(
     val name:String,
-    val totalValue:Int,
-    val totalSource:Int,
+    val totalValue: Double,
+    val totalSource:Double,
     val idMap:Int
 )
